Frameworks/Sovereign-C2 (Enhanced) Under Testing/c2_server/requirements.txt
fastapi
uvicorn
typer
pycryptodome
pynput
requests
paramiko
